Double tuneable should be double range.

+++ Xcode/SmartSensors/src/tuneable.cpp
@@ -52,7 +52,7 @@ void registerTuneable(int& value, int min, int max, const string& description) {
     ((ofApp *) ofGetAppPtr())->registerTuneable(t);
 }
 
-void registerTuneable(double& value, int min, int max, const string& description) {
+void registerTuneable(double& value, double min, double max, const string& description) {
     void* address = &value;
     if (allTuneables.find(address) != allTuneables.end()) {
         return;
